+++ package.json
@@ -15,6 +15,7 @@
     "@babel/preset-env": "^7.9.6",
     "@babel/preset-typescript": "^7.8.3",
     "@babel/runtime": "^7.9.6",
+    "@std/esm": "^0.26.0",
     "@types/chai": "^4.2.9",
     "@types/gl-vec2": "^1.3.0",
     "@types/lodash": "^4.14.150",
@@ -50,6 +51,7 @@
     "sinon": "5.1.1",
     "sinon-chai": "^3.5.0",
     "source-map-loader": "^0.2.4",
+    "ts-mocha": "^7.0.0",
     "typescript": "^3.8.2",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11",
@@ -60,9 +62,16 @@
   },
   "scripts": {
     "check-types": "tsc",
-    "test": "npx grunt test",
+    "test:browser-specific": "need something to run a browser suite here",
+    "test:browser-universal": "need something to run a browser suite here",
+    "test:import": "npx mocha -r @std/esm test/test-import.mjs",
+    "test:node-specific": "npx ts-mocha -p test/tsconfig.json src/**/test/node/*.spec.*",
+    "test:node-universal": "npx ts-mocha -p test/tsconfig.json -r chai/register-expect src/**/test/*.spec.*",
     "test:require": "npx mocha test/test-require",
-    "test:import": "npx mocha -r @std/esm -r ts-node/register test/test-import.mjs",
+    "test:browser-all": "npm run test:browser-universal && npm run test:browser-specific",
+    "test:module": "npm run test:require && npm run test:import",
+    "test:node-all": "npm run test:node-universal && npm run test:node-specific",
+    "test": "npm run test:module && npm run test:node-all && npx grunt test",
     "integrationtest": "grunt integrationtest",
     "build:dev": "npx cross-env BUILD_ENV=development webpack --config configs/webpack.config.js",
     "build:prod": "npx cross-env BUILD_ENV=production webpack --config configs/webpack.config.min.js",
